+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * InputString and Sample have been moved from stamina-induction to stamina-core as ADL
   and Walking rely on them.
+* Walking methods `parses?`, `accepts?` and `rejects?` are now aliased as `parse?`,
+  `accept?` and `reject?`, respectively.
 
 # 0.5.3 / 2012-02-25